requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
mangum>=0.17.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
pydantic-settings>=2.3.0
python-multipart>=0.0.9
httpx>=0.27.0
email-validator>=2.1.0
